requirements.txt
# Core
Django>=2.2,<2.3
djangorestframework==3.9.4

# Tahoe requirements
# TODO: Replace with tahoe-sites and edx-organizations==5.2.0 (https://github.com/appsembler/tahoe-figures-plugins/issues/3)
https://github.com/appsembler/edx-organizations/archive/5.2.0-appsembler14.tar.gz
django-model-utils<=4.0.0

# Testing
pytest>=6.1,<=6.2.4
pytest-cov>=2.12.1,<=3.0.0
pytest-django==4.4.0
coverage==5.5
